theme utils refactor
