--- src/utils/getSystemTheme.js
@@ -1,20 +1,0 @@
-/**
- * Copyright (c) Hathor Labs and its affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-/**
- * Determines the system's preferred color scheme.
- *
- * @returns {string} The system's preferred theme, either 'dark' or 'light'.
- */
-const getSystemTheme = () => {
-  if (window.matchMedia && window.matchMedia('(prefers-color-scheme: dark)').matches) {
-    return 'dark';
-  }
-  return 'light';
-};
-
-export default getSystemTheme;
